git-svn-id: https://svn.code.sf.net/p/oo-open/code@104 2a060ecb-2c39-0410-a787-992002912dfc

+++ OblivionOnline/OOPackets.h
@@ -69,7 +69,7 @@ struct OOPkgDisconnect
 {
 	OOPacketType etypeID;
 	short Flags;
-	short PlayerID;
+	int PlayerID;
 };
 
 /*
